// admin_live_mall/src/pages/Data/ProductData/ProductOverview/data.d.ts
export interface OverviewGoodsDataType {
  createId: number;
  createTime: string;
  id: number;
  isDelete: boolean;
  modifyId: number;
  modifyTime: string;
  num: number;
  onSell: string;
  pv: number;
  shopId: number;
  tenantId: number;
  uv: number;
  addCartCount: number;
  placeCount: number;
  payCount: number;
}